Add files via upload
This project involved analysing access to water in the
national, urban and rural populations of 213 countries in
2020 from the United Nations Estimates on Water Use
dataset. The access to water was categorised into four
levels of quality–basic, limited, unimproved and
surface. At the national level, about two-thirds of the
populations had 75% and above access to basic level
water services.